pyproject.toml
[project]
name = "cb-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.7.2",
 "alembic>=1.16.1",
 "anytree>=2.13.0",
 "bcrypt>=4.3.0",
 "cryptography>=45.0.3",
 "fastapi>=0.115.12",
 "httpx>=0.28.1",
 "passlib[bcrypt]>=1.7.4",
 "peewee>=3.18.1",
 "pydantic>=2.11.4",
 "pydantic-settings>=2.9.1",
 "pytest-asyncio>=0.21.0",
 "python-dotenv>=1.1.0",
 "python-jose[cryptography]>=3.5.0",
 "python-multipart>=0.0.20",
 "requests>=2.32.3",
 "slowapi>=0.1.9",
 "tomlkit>=0.13.3",
 "tortoise-orm>=0.20.0",
 "uvicorn>=0.34.2",
]

[dependency-groups]
dev = ["ty>=0.0.1a2"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.aerich]
tortoise_orm = "src.infra.tortoise_client.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
